refactor: modularize SmartChat component
