+++ frontend/components/search-wizard/modes/smart-chat/constants.ts
@@ -1,0 +1,5 @@
+export const INTRO_MESSAGE = [
+  "Hi! I'm here to help you find the right commercial property.",
+  "Tell me what you're looking for — be as detailed or brief as you want.",
+  "For example: 'I need a 100k sqft industrial warehouse with 32ft clear height in Chicago for lease, with at least 20 dock doors.'",
+].join(" ");
