+++ frontend/components/search-wizard/modes/smart-chat/types.ts
@@ -1,0 +1,7 @@
+export type ChatRole = "user" | "assistant";
+
+export type ChatMessage = {
+  id: string;
+  role: ChatRole;
+  content: string;
+};
